+++ onadata/apps/viewer/models/parsed_instance.py
@@ -25,15 +25,12 @@ from onadata.libs.utils.common_tags import ID, UUID, ATTACHMENTS, GEOLOCATION,\
 from onadata.libs.utils.osm import save_osm_data_async
 from onadata.libs.utils.model_tools import queryset_iterator
 from onadata.libs.utils.mongo import _is_invalid_for_mongo
+from onadata.apps.viewer.parsed_instance_tools import get_where_clause
+from onadata.apps.viewer.parsed_instance_tools import NONE_JSON_FIELDS
 
 ASYNC_POST_SUBMISSION_PROCESSING_ENABLED = \
     getattr(settings, 'ASYNC_POST_SUBMISSION_PROCESSING_ENABLED', False)
 DATETIME_FORMAT = '%Y-%m-%dT%H:%M:%S'
-KNOWN_DATES = ['_submission_time']
-NONE_JSON_FIELDS = {
-    '_submission_time': 'date_created',
-    '_id': 'id'
-}
 
 
 class ParseError(Exception):
@@ -50,17 +47,6 @@ def datetime_from_str(text):
     except Exception:
         return None
     return dt
-
-
-def _json_sql_str(key, known_integers=[], known_dates=[]):
-    _json_str = u"json->>%s"
-
-    if key in known_integers:
-        _json_str = u"CAST(json->>%s AS INT)"
-    elif key in known_dates:
-        _json_str = u"CAST(json->>%s AS TIMESTAMP)"
-
-    return _json_str
 
 
 def dict_for_mongo(d):
@@ -89,45 +75,6 @@ def get_name_from_survey_element(element):
 def _parse_sort_fields(fields):
     for field in fields:
         yield NONE_JSON_FIELDS.get(field, field)
-
-
-def _parse_where(query, known_integers, or_where, or_params):
-    # using a dictionary here just incase we will need to filter using
-    # other table columns
-    where, where_params = [], []
-    OPERANDS = {
-        '$gt': '>',
-        '$gte': '>=',
-        '$lt': '<',
-        '$lte': '<=',
-        '$i': '~*'
-    }
-    for field_key, field_value in query.iteritems():
-        if isinstance(field_value, dict):
-            if field_key in NONE_JSON_FIELDS:
-                json_str = NONE_JSON_FIELDS.get(field_key)
-            else:
-                json_str = _json_sql_str(
-                    field_key, known_integers, KNOWN_DATES)
-            for key, value in field_value.iteritems():
-                _v = None
-                if key in OPERANDS:
-                    where.append(
-                        u' '.join([json_str, OPERANDS.get(key), u'%s'])
-                    )
-                _v = value
-                if field_key in KNOWN_DATES:
-                    _v = datetime.datetime.strptime(
-                        _v[:19], MONGO_STRFTIME)
-                if field_key in NONE_JSON_FIELDS:
-                    where_params.extend([unicode(_v)])
-                else:
-                    where_params.extend((field_key, unicode(_v)))
-        else:
-            where.append(u"json->>%s = %s")
-            where_params.extend((field_key, unicode(field_value)))
-
-    return where + or_where, where_params + or_params
 
 
 def _query_iterator(sql, fields=None, params=[], count=False):
@@ -169,41 +116,6 @@ def get_etag_hash_from_query(queryset, sql=None, params=None):
             return etag_hash[0]
 
     return u'%s' % datetime.datetime.utcnow()
-
-
-def get_where_clause(query, form_integer_fields=[]):
-    known_integers = ['_id'] + form_integer_fields
-    where = []
-    where_params = []
-
-    try:
-        if query and isinstance(query, six.string_types):
-            query = json.loads(query)
-            or_where = []
-            or_params = []
-            if isinstance(query, list):
-                query = query[0]
-
-            if '$or' in query.keys():
-                or_dict = query.pop('$or')
-                for l in or_dict:
-                    or_where.extend([u"json->>%s = %s" for i in l.items()])
-                    [or_params.extend(i) for i in l.items()]
-
-                or_where = [u"".join([u"(", u" OR ".join(or_where), u")"])]
-
-            where, where_params = _parse_where(query, known_integers,
-                                               or_where, or_params)
-
-    except (ValueError, AttributeError) as e:
-        if query and isinstance(query, six.string_types) and \
-                query.startswith('{'):
-            raise e
-        # cast query param to text
-        where = [u"json::text ~* cast(%s as text)"]
-        where_params = [query]
-
-    return where, where_params
 
 
 def _start_index_limit(records, sql, fields, params, sort, start_index, limit):
